объединила экраны вопросов в один js модуль
