+++ js/screens/question-screen-view.js
@@ -3,9 +3,10 @@ import getScreenMarkup from '../markup/get-question-markup';
 import getTaskMarkup from '../markup/get-task-markup';
 import getAnswersMarkup from '../markup/get-answers-markup';
 
-export default class ArtistScreenView extends AbstractView {
+export default class QuestionScreenView extends AbstractView {
   constructor(type, options, link, question, mistakes, time) {
     super();
+
     this._type = type;
     this._options = options;
     this._link = link;
@@ -15,6 +16,12 @@ export default class ArtistScreenView extends AbstractView {
 
     this._answers = getAnswersMarkup(this._type, this._options);
     this._task = getTaskMarkup(this._type, this._question, this._answers, this._link);
+
+    // только у жанра
+    if (this._type === `genre`) {
+      this.answerBtn = this.element.querySelector(`.genre-answer-send`);
+      this.checkboxes = [...this.element.querySelectorAll(`input[type=checkbox]`)];
+    }
   }
 
   get template() {
@@ -25,14 +32,34 @@ export default class ArtistScreenView extends AbstractView {
     this.minutes = this.element.querySelector(`.timer-value-mins`);
     this.seconds = this.element.querySelector(`.timer-value-secs`);
 
-    const radioButtons = this.element.querySelector(`.main-list`);
-    radioButtons.addEventListener(`click`, (evt) => {
-      evt.preventDefault();
-      this.onRadioBtnClick(evt);
-    });
+    if (this._type === `genre`) {
+      const answerBtn = this.element.querySelector(`.genre-answer-send`);
+      answerBtn.addEventListener(`click`, (evt) => {
+        evt.preventDefault();
+        this.onSendAnswerBtnClick(evt);
+      });
+
+      const checkboxes = [...this.element.querySelectorAll(`input[type=checkbox]`)];
+      checkboxes.forEach((checkbox) => {
+
+        checkbox.addEventListener(`click`, () => {
+          this.onAnswerClick();
+        });
+      });
+    } else {
+      const radioButtons = this.element.querySelector(`.main-list`);
+      radioButtons.addEventListener(`click`, (evt) => {
+        evt.preventDefault();
+        this.onSendAnswerBtnClick(evt);
+      });
+    }
   }
 
-  onRadioBtnClick() {
+  onAnswerClick() {
+
+  }
+
+  onSendAnswerBtnClick() {
 
   }
